feat(stage): 将多选逻辑封装到StageMultiDragResize
